Create utils.h
